server/requirements.txt
openai>=1.0.0
uvicorn>=0.24.0
fastapi>=0.104.0
pydantic>=2.0.0
openenv-core>=0.2.2
websockets>=12.0
